fix: do not pass sanitized directories into utils.dirs_match
The `utils.dirs_match()` function is not expecting input directories
to already be sanitized and can result in failure to properly detect
matches with directories.

The `init.allowed_dir()` function is sanitizing the current working
directory before passing as input to `utils.dirs_match()`, causing
failure to detect match with the following setup on a macOS system:

```lua
allowed_dirs = {
  '~/.config',
}
```

And invoking `nvim` while in `~/.config/nvim` working directory.

This change removes `init.cwd()` which is only ever used twice, one of
which is incorrectly in `init.allowed_dir()`, the other in
`init.current()`. This updates `init.current()` to sanitize the working
directory as is needed, and updates `init.allowed_dir()` to use the
unsanitized working directory.